【リファクタ】OAuth認可はcommonでの処理として扱うようにした
