+++ src/accessTokenInitialize.ts
@@ -1,56 +1,11 @@
 import axios from 'axios';
 import fs from 'fs';
 import emoji from 'node-emoji';
-import open from 'open';
 import { Answers, prompt, QuestionCollection } from 'inquirer';
 import sleep from 'sleep-promise';
-import { User } from '@/types/qiita';
 import { initializeAndLoadQiitaDir } from './commons/qiitaSettings';
-import { loadAuthenticatedUser, getAccessToken } from './commons/qiitaApis';
-import { createServer, Server, ServerResponse, IncomingMessage } from 'http';
-import { AddressInfo } from 'net';
-import enableDestroy from 'server-destroy';
-import { ReadonlyDeep } from 'type-fest';
-import crypto from 'crypto';
-
-const globalOauth2ClientSettings = {
-  clientId: '7aca58a519a2d1cb83d8117d5d7a8210a3c2cd53',
-  clientSecret: 'cb1d5595b9122742fc0e56d16610f2a93a3a8c11',
-};
-
-// Oauth認可を行う時に受け取るサーバのポート番号
-const serverPortNumber = 59116;
-
-async function startLocalServer(): Promise<Server> {
-  return new Promise<Server>((resolve) => {
-    const server = createServer();
-    enableDestroy(server);
-    server.listen(serverPortNumber, () => resolve(server));
-  });
-}
-
-async function recieveOauthCallbackCode(server: Server): Promise<string> {
-  return new Promise<string>((resolve, reject) => {
-    server.on(
-      'request',
-      (
-        request: ReadonlyDeep<IncomingMessage>,
-        response: ReadonlyDeep<ServerResponse>
-      ) => {
-        const urlParts = new URL(request.url ?? '', 'http://localhost')
-          .searchParams;
-        const code = urlParts.get('code');
-        const error = urlParts.get('error');
-        if (code) {
-          resolve(code);
-        } else {
-          reject(error);
-        }
-        response.end('Logged in! You may close this page. ');
-      }
-    );
-  });
-}
+import { loadAuthenticatedUser } from './commons/qiitaApis';
+import { oauthLogin } from './commons/accesstokenManager';
 
 export async function accessTokenInitialize(): Promise<number> {
   try {
@@ -82,30 +37,7 @@ export async function accessTokenInitialize(): Promise<number> {
     await sleep(1500);
     // 1.5 seconds later
     // open qiita admin by web default web browser
-    // OAuth認可を行う時に
-    const authorizeState = crypto.randomBytes(12).toString('hex');
-    const permissionScope = [
-      'read_qiita',
-      'write_qiita',
-      'read_qiita_team',
-      'write_qiita_team',
-    ].join('+');
-    const qiitaAuthorizeUrl = `https://qiita.com/api/v2/oauth/authorize?client_id=${globalOauth2ClientSettings.clientId}&scope=${permissionScope}&state=${authorizeState}`;
-    const openBrowserPromise = open(qiitaAuthorizeUrl);
     //await open('https://qiita.com/settings/applications');
-
-    const server = await startLocalServer();
-    // { address: '::', family: 'IPv6', port: 50441 } のような形でportの値のみを取得する
-    const { port } = server.address() as AddressInfo;
-    await openBrowserPromise;
-    const authCode = await recieveOauthCallbackCode(server).finally(() => {
-      server.destroy();
-    });
-    const accessTokenResponse = await getAccessToken(
-      globalOauth2ClientSettings.clientId,
-      globalOauth2ClientSettings.clientSecret,
-      authCode
-    );
 
     /*
     // ユーザ入出力形式指定
@@ -126,7 +58,7 @@ export async function accessTokenInitialize(): Promise<number> {
     // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
     const token: string = answers.token;
     */
-    const token = accessTokenResponse.data.token;
+    const token = await oauthLogin();
     const res = await loadAuthenticatedUser(token);
     const qiitaUser = {
       id: res.data.id,
